docs(gallery): split plot_gradient_flow_field into 5 flow examples
Quantum flow dynamics, radial burst, DNA helix, Fibonacci spiral
galaxy, and Voronoi tessellation now appear as separate gallery
cards instead of a single entry that bundled five unrelated figures.
